[ssci] Added Shipped field to READMEs
This CL adds the Shipped field (and may update the
License File field) in Chromium READMEs. Changes were
automatically created, so if you disagree with any of
them (e.g. a package is used only for testing purposes
and is not shipped), comment the suggested change and
why.

See the LSC doc at go/lsc-chrome-metadata.

Bug: b:285450740
Change-Id: Iaf9c0531c3a470d6bb5c8b5b51defa4659c83d57
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/4666404
Reviewed-by: Frank Tang <ftang@chromium.org>