[fuchsia] Add slots to add or remove compile configs
A while back we added icu_fuchsia_extra_compile_flags support to ICU,
allowing us to add new compile flags to ICU compilation in Fuchsia
instantly, without having to wait for an ICU roll.

This change is a more generic version of that same thing: it allows us
to inject or remove special GN configs, under the same constraints.

Bug: b/328487023
Change-Id: If6022b5319b73730b941f51e02c96cf1f16384c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/deps/icu/+/5348254
Reviewed-by: Frank Tang <ftang@chromium.org>
Reviewed-by: David Yeung <dayeung@chromium.org>